refactor(unix): split test_support with FakeBlockstore
